[refactor] 피드백 반영 스타일링
